adjusted functionalities, added placeholders for new ones

+++ xlsx_to_ary.py
@@ -25,19 +25,39 @@ def load_fields(filename):
     for i in range(2,sheet.max_row+1):
         if sheet['A'+str(i)].value is not None:
             print(sheet['A'+str(i)].value)
-            dict_list.append({'word':sheet['A'+str(i)].value,'plural':sheet['B'+str(i)].value,'lang_origin':sheet['C'+str(i)].value,'word_origin':sheet['D'+str(i)].value
-                            ,'transliteration_origin':sheet['E'+str(i)].value,'meaning_origin':sheet['F'+str(i)].value,'certainty_origin':sheet['G'+str(i)].value
-                            ,'second_lang_origin':sheet['H'+str(i)].value,'second_word_origin':sheet['I'+str(i)].value,'second_transl_origin':sheet['J'+str(i)].value
-                            ,'second_meaning_origin':sheet['K'+str(i)].value,'origins_relationship':sheet['L'+str(i)].value,'second_cert_origin':sheet['N'+str(i)].value
-                            ,'part_of_speech':sheet['M'+str(i)].value,'pronunciation':sheet['O'+str(i)].value,'meaning':sheet['P'+str(i)].value
-                            ,'meaning_example':sheet['Q'+str(i)].value,'figurative_meaning':sheet['R'+str(i)].value,'figurative_example':sheet['S'+str(i)].value
-                            ,'synonyms':sheet['T'+str(i)].value,'dialectal_variations':sheet['U'+str(i)].value,'related':sheet['V'+str(i)].value
-                            ,'expressions':sheet['W'+str(i)].value,'source':sheet['X'+str(i)].value,'picture':sheet['Y'+str(i)].value
-                            ,'picture_comment':sheet['Z'+str(i)].value})
+            dict_list.append({'word':sheet['A'+str(i)].value
+                             ,'plural':sheet['B'+str(i)].value
+                             ,'feminine':sheet['C'+str(i)].value
+                             ,'feminine_plural':sheet['D'+str(i)].value
+                             ,'lang_origin':sheet['E'+str(i)].value
+                             ,'word_origin':sheet['F'+str(i)].value
+                             ,'transliteration_origin':sheet['G'+str(i)].value
+                             ,'meaning_origin':sheet['H'+str(i)].value
+                             ,'certainty_origin':sheet['I'+str(i)].value
+                             ,'second_lang_origin':sheet['J'+str(i)].value
+                             ,'second_word_origin':sheet['K'+str(i)].value
+                             ,'second_transl_origin':sheet['L'+str(i)].value
+                             ,'second_meaning_origin':sheet['M'+str(i)].value
+                             ,'second_cert_origin':sheet['N'+str(i)].value
+                             ,'origins_relationship':sheet['O'+str(i)].value
+                             ,'part_of_speech':sheet['P'+str(i)].value
+                             ,'pronunciation':sheet['Q'+str(i)].value
+                             ,'meaning':sheet['R'+str(i)].value
+                             ,'meaning_example':sheet['S'+str(i)].value
+                             ,'figurative_meaning':sheet['T'+str(i)].value
+                             ,'figurative_example':sheet['U'+str(i)].value
+                             ,'synonyms':sheet['V'+str(i)].value
+                             ,'antonyms':sheet['W'+str(i)].value
+                             ,'dialectal_variations':sheet['X'+str(i)].value
+                             ,'related':sheet['Y'+str(i)].value
+                             ,'expressions':sheet['Z'+str(i)].value
+                             ,'source':sheet['AA'+str(i)].value
+                             ,'picture':sheet['AB'+str(i)].value
+                             ,'picture_comment':sheet['AC'+str(i)].value})
 
     return dict_list
 
-def get_origin_code(lang_origin,word_origin,transliteration_origin,pronunciation_origin,meaning_origin
+def get_origin_code(lang_origin,word_origin,transliteration_origin,meaning_origin
                                                  ,certainty_origin,second_lang_origin,second_word_origin
                                                  ,second_transl_origin,second_meaning_origin,second_cert_origin
                                                  ,origins_relationship):
@@ -61,7 +81,7 @@ def get_origin_code(lang_origin,word_origin,transliteration_origin,pronunciation
                     code+=" ؤلا "
                 if second_cert_origin == "unsure":
                     code+="واقيلا"
-                code+=" من {{Wt/ary/étyl|"+get_lang_code(second_lang_origin)+"|ary|"
+                code+=" من {{"+"Wt/ary/étyl|{}|ary|".format(get_lang_code(second_lang_origin))
                 if second_word_origin is not None:
                     code+=second_word_origin
                 if second_transl_origin is not None:
@@ -76,11 +96,30 @@ def get_origin_code(lang_origin,word_origin,transliteration_origin,pronunciation
     
      
 
-def get_main_section_code(word,part_of_speech,plural,pronunciation,meaning,meaning_example):
-    code = "=== {{Wt/ary/S|"+translate_part_of_speech(part_of_speech)+"|ary}} ===\n"
-    if plural is not None:
-        code+="{{Wt/ary/فرد-جمع1|s=Wt/ary/"+word+"|p=Wt/ary/"+plural+"}}"
-    code+="\n'''"+word+"''' {{Wt/ary/pron|"+pronunciation+"|ary}}" #+meaning
+def get_main_section_code(word,part_of_speech,plural,pronunciation,meaning,meaning_example,feminine,feminine_plural):
+    code = "=== {"+"{{Wt/ary/S|{}|ary}}".format(translate_part_of_speech(part_of_speech)) +"} ===\n"
+    if feminine is None:
+        if plural is not None:
+            plurals = [w.strip() for w in plural.split('..')]
+            if len(plurals) == 1:
+                code+="{"+"{{Wt/ary/فرد-جمع1|s=Wt/ary/{}|p=Wt/ary/{}}}".format(word,plurals[0])+"}"
+            else:
+                code+="{"+"{{Wt/ary/فرد-جمع1|s=Wt/ary/{}|p=Wt/ary/{}|p2=Wt/ary/{}}}".format(word,plurals[0],plurals[1])+"}"
+    else:
+        if plural is not None:
+            plurals = [w.strip() for w in plural.split('..')]
+            code+="{"+"{{Wt/ary/فرد-جمع2| د.ف=Wt/ary/{}| د.ج=Wt/ary/{}| ن.ف=Wt/ary/{}| ن.ج=Wt/ary/{}}}".format(word,plurals[0],feminine,feminine_plural)+"}"
+        else:
+            code+="{"+"{{Wt/ary/فرد-جمع2| د.ف=Wt/ary/{}| ن.ف=Wt/ary/{}}}".format(word,plural[0],feminine,feminine_plural)+"}"
+    code+="\n'''{}''' ".format(word)
+    pronunciations = [w.strip() for w in pronunciation.split('..')]
+    if len(pronunciations) == 1:
+        code+="{"+"{{Wt/ary/pron|{}|ary}}".format(pronunciation)+"}" #+meaning
+    elif len(pronunciations) > 1:
+        code+="{"+"{{Wt/ary/pron|{}|ary}}".format(pronunciations[0])+"}"
+        for i in range(1,len(pronunciations)):
+            code+=" ؤلا {"+"{{Wt/ary/pron|{}|ary}}".format(pronunciations[i])+"}"
+            
     meaning = [m.strip() for m in meaning.split('..')]
     if meaning_example is not None:
         meaning_example = [me.strip() for me in meaning_example.split('..')]
@@ -121,7 +160,19 @@ def get_synonyms_code(synonyms):
         code = "==== {{Wt/ary/S|مرادفات}} ====\n"
         
         for synonym in synonyms:
-            code +="*[[Wt/ary/"+synonym+"|"+synonym+"]]\n"
+            code +="*[[Wt/ary/{}|{}]]\n".format(synonym,synonym)
+        return code
+
+    return ''
+
+def get_antonyms_code(antonyms):
+    
+    if antonyms is not None:
+        antonyms = [w.strip() for w in antonyms.split('.')]
+        code = "==== {{Wt/ary/S|معاكسات}} ====\n"
+        
+        for antonym in antonyms:
+            code +="*[[Wt/ary/{}|{}]]\n".format(antonym,antonym)
         return code
 
     return ''
@@ -131,7 +182,7 @@ def get_related_code(related):
         related = [w.strip() for w in related.split('.')]
         code = "==== {{Wt/ary/S|مشتقات}} ====\n" #{{wt/ary/(}}\n"
         for r in related:
-            code +="*[[Wt/ary/"+r+"|"+r+"]]\n"
+            code +="*[[Wt/ary/{}|{}]]\n".format(r,r)
         #code+="{{wt/ary/(}}"
         return code
 
@@ -144,17 +195,18 @@ def get_dialect_var_code(dialectal_variations):
         code = "==== {{Wt/ary/S|شكال لهجاوية}} ====\n"
         
         for dialectal_variation in dialectal_variations:
-            code +="*[[|Wt/ary/"+dialectal_variation+"|"+dialectal_variation+"]]\n"
+            code +="*[[|Wt/ary/{}|{}]]\n".format(dialectal_variation,dialectal_variation)
         return code
 
     return ''
 
 def get_image_code(picture,picture_comment,word):
+    
     if picture is not None:
-        code = "[[File:"+picture+"|thumb|left|upright|"+word #+"]]"
-        if picture_comment is not None:
-            code+=". "+picture_comment
-        code+="]]"
+        if picture_comment is None:
+            picture_comment = word
+            
+        code = "[[File:{}|thumb|left|upright|{}".format(picture,picture_comment)+"]]"
         return code
     return ''
 
@@ -163,7 +215,7 @@ def get_expressions_code(expressions):
         expressions = [ex.strip() for ex in expressions.split('..')]
         code = "==== {{Wt/ary/S|تعبيرات}} ===="
         for ex in expressions:
-            code+="\n*[[|Wt/ary/"+ex+"|"+ex+"]]"
+            code+="\n*[[|Wt/ary/{}|{}]]".format(ex,ex)
         return code
 
     return ''
@@ -178,14 +230,28 @@ def get_source_link_code(source):
     return ''
 
 def export_files(full_code,export_filename):
-    with open(export_filename,'w',encoding='utf-8') as f:
-        f.write(full_code)
+    if not os.path.exists(export_filename):
+        full_code = "== {{Wt/ary/langue|ary}} ==\n"+full_code
+        with open(export_filename,'w',encoding='utf-8') as f:
+            f.write(full_code)
+    else:
+        with open(export_filename,'a',encoding='utf-8') as f:
+            f.write('\n\n'+full_code)
 
 def rnvl(code):
     if code is None or code.strip() == '':
         return ''
     else:
         return code
+
+def get_plural_full_code(word,part_of_speech,plural,plural_pronunciation,feminine,feminine_plural):
+    pass
+
+def get_feminine_full_code(word,part_of_speech,plural,feminine_pronunciation,feminine,feminine_plural):
+    pass
+
+def get_fem_plural_full_code(word,part_of_speech,plural,fem_plural_pronunciation,feminine,feminine_plural):
+    pass
     
 
 
@@ -203,6 +269,8 @@ if __name__=='__main__':
         if word is None or part_of_speech is None or meaning is None:
             continue
         plural             = dict_list[i]['plural']
+        feminine           = dict_list[i]['feminine']
+        feminine_plural    = dict_list[i]['feminine_plural']
         pronunciation      = dict_list[i]['pronunciation']
         meaning_example    = dict_list[i]['meaning_example']
         figurative_meaning = dict_list[i]['figurative_meaning']
@@ -211,12 +279,12 @@ if __name__=='__main__':
         lang_origin        = dict_list[i]['lang_origin']
         word_origin        = dict_list[i]['word_origin']
         synonyms           = dict_list[i]['synonyms']
+        antonyms           = dict_list[i]['antonyms']
         dialectal_variations = dict_list[i]['dialectal_variations']
         related            = dict_list[i]['related']
         expressions        = dict_list[i]['expressions']
         picture            = dict_list[i]['picture']
         transliteration_origin = dict_list[i]['transliteration_origin']
-        pronunciation_origin   = dict_list[i]['pronunciation_origin']
         meaning_origin     = dict_list[i]['meaning_origin']
         certainty_origin   = dict_list[i]['certainty_origin']
         second_lang_origin = dict_list[i]['second_lang_origin']
@@ -228,12 +296,13 @@ if __name__=='__main__':
         picture_comment      = dict_list[i]['picture_comment']
         
 
-        main_section = get_main_section_code(word,part_of_speech,plural,pronunciation,meaning,meaning_example) #done
-        ori_section = get_origin_code(lang_origin,word_origin,transliteration_origin,pronunciation_origin,meaning_origin
+        main_section = get_main_section_code(word,part_of_speech,plural,pronunciation,meaning,meaning_example,feminine,feminine_plural) #done
+        ori_section = get_origin_code(lang_origin,word_origin,transliteration_origin,meaning_origin
                                                  ,certainty_origin,second_lang_origin,second_word_origin
                                                  ,second_transl_origin,second_meaning_origin,second_cert_origin
                                                  ,origins_relationship)
         syn_section = get_synonyms_code(synonyms)
+        ant_section = get_antonyms_code(antonyms)
         dia_section = get_dialect_var_code(dialectal_variations)
         rel_section = get_related_code(related)
         fig_section = get_figurative_code(figurative_meaning,figurative_example)
@@ -241,8 +310,10 @@ if __name__=='__main__':
         exp_section = get_expressions_code(expressions)
         src_section = get_source_link_code(source)
         #print(main_section)
-        full_code = "== {{Wt/ary/langue|ary}} =="+rnvl('\n'+ori_section+'\n')+rnvl('\n'+im_section+'\n')+'\n'+main_section+'\n'+rnvl('\n'+fig_section+'\n')
-        full_code +=rnvl('\n'+syn_section+'\n')+rnvl('\n'+dia_section+'\n')+rnvl('\n'+rel_section+'\n')+rnvl('\n'+exp_section)+rnvl('\n'+src_section)
+        full_code = rnvl('\n'+ori_section+'\n')+rnvl('\n'+im_section+'\n')+'\n'+main_section+rnvl('\n'+fig_section+'\n')
+        full_code +=rnvl('\n'+syn_section+'\n')+rnvl('\n'+ant_section+'\n')+rnvl('\n'+dia_section+'\n')+rnvl('\n'+rel_section+'\n')
+        full_code +=rnvl('\n'+exp_section)+rnvl('\n'+src_section)
+        
         export_files(full_code.strip(),EXPORT_FOLDER+word+'.txt')
